requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pyodbc>=5.0.0
pandas>=2.0.0
pyyaml>=6.0
